type casted the inputs to decoder to bfloat16

+++ models/official/projects/detr/modeling/transformer.py
@@ -853,5 +853,5 @@ class TransformerDecoderBlock(tf.keras.layers.Layer):
     if self._norm_first:
       layer_output = source_attention_output + layer_output
     else:
-        layer_output = self.output_layer_norm(layer_output + attention_output)
+        layer_output = self.output_layer_norm(tf.cast(layer_output, tf.float32) + attention_output)
     return layer_output, cache
